refactoring and logic of NamingConvention
