+++ packages/core/src/factories/JsonMetadataFactory.ts
@@ -3,9 +3,9 @@ import ts from "typescript";
 import { TransformerError } from "../context/TransformerError";
 import { AtomicPredicator } from "../programmers/helpers/AtomicPredicator";
 import { MetadataSchema } from "../schemas/metadata/MetadataSchema";
+import { MetadataStorage } from "../schemas/metadata/MetadataStorage";
 import { ValidationPipe } from "../typings/ValidationPipe";
 import { MetadataFactory } from "./MetadataFactory";
-import { MetadataStorage } from "./MetadataStorage";
 
 export namespace JsonMetadataFactory {
   export interface IProps {
