+++ packages/core/src/factories/MetadataFactory.ts
@@ -7,10 +7,10 @@ import { MetadataFunction } from "../schemas/metadata/MetadataFunction";
 import { MetadataObject } from "../schemas/metadata/MetadataObject";
 import { MetadataObjectType } from "../schemas/metadata/MetadataObjectType";
 import { MetadataSchema } from "../schemas/metadata/MetadataSchema";
+import { MetadataStorage } from "../schemas/metadata/MetadataStorage";
 import { MetadataTupleType } from "../schemas/metadata/MetadataTupleType";
 import { ValidationPipe } from "../typings/ValidationPipe";
 import { ExpressionFactory } from "./ExpressionFactory";
-import { MetadataStorage } from "./MetadataStorage";
 import { explore_metadata } from "./internal/metadata/explore_metadata";
 import { iterate_metadata_collection } from "./internal/metadata/iterate_metadata_collection";
 import { iterate_metadata_sort } from "./internal/metadata/iterate_metadata_sort";
