+++ packages/core/src/programmers/RandomProgrammer.ts
@@ -9,7 +9,6 @@ import { IdentifierFactory } from "../factories/IdentifierFactory";
 import { LiteralFactory } from "../factories/LiteralFactory";
 import { MetadataCommentTagFactory } from "../factories/MetadataCommentTagFactory";
 import { MetadataFactory } from "../factories/MetadataFactory";
-import { MetadataStorage } from "../factories/MetadataStorage";
 import { StatementFactory } from "../factories/StatementFactory";
 import { TemplateFactory } from "../factories/TemplateFactory";
 import { TypeFactory } from "../factories/TypeFactory";
@@ -20,6 +19,7 @@ import { MetadataMap } from "../schemas/metadata/MetadataMap";
 import { MetadataObjectType } from "../schemas/metadata/MetadataObjectType";
 import { MetadataSchema } from "../schemas/metadata/MetadataSchema";
 import { MetadataSet } from "../schemas/metadata/MetadataSet";
+import { MetadataStorage } from "../schemas/metadata/MetadataStorage";
 import { MetadataTemplate } from "../schemas/metadata/MetadataTemplate";
 import { MetadataTuple } from "../schemas/metadata/MetadataTuple";
 import { MetadataTupleType } from "../schemas/metadata/MetadataTupleType";
