+++ packages/core/src/programmers/http/HttpFormDataProgrammer.ts
@@ -7,13 +7,13 @@ import { ITypiaContext } from "../../context/ITypiaContext";
 import { TransformerError } from "../../context/TransformerError";
 import { IdentifierFactory } from "../../factories/IdentifierFactory";
 import { MetadataFactory } from "../../factories/MetadataFactory";
-import { MetadataStorage } from "../../factories/MetadataStorage";
 import { StatementFactory } from "../../factories/StatementFactory";
 import { TypeFactory } from "../../factories/TypeFactory";
 import { MetadataArrayType } from "../../schemas/metadata/MetadataArrayType";
 import { MetadataObjectType } from "../../schemas/metadata/MetadataObjectType";
 import { MetadataProperty } from "../../schemas/metadata/MetadataProperty";
 import { MetadataSchema } from "../../schemas/metadata/MetadataSchema";
+import { MetadataStorage } from "../../schemas/metadata/MetadataStorage";
 import { FunctionProgrammer } from "../helpers/FunctionProgrammer";
 import { HttpMetadataUtil } from "../helpers/HttpMetadataUtil";
 import { FeatureProgrammer } from "../internal/FeatureProgrammer";
