+++ packages/core/src/programmers/misc/MiscCloneProgrammer.ts
@@ -6,7 +6,6 @@ import { TransformerError } from "../../context/TransformerError";
 import { ExpressionFactory } from "../../factories/ExpressionFactory";
 import { IdentifierFactory } from "../../factories/IdentifierFactory";
 import { MetadataFactory } from "../../factories/MetadataFactory";
-import { MetadataStorage } from "../../factories/MetadataStorage";
 import { StatementFactory } from "../../factories/StatementFactory";
 import { TypeFactory } from "../../factories/TypeFactory";
 import { MetadataArray } from "../../schemas/metadata/MetadataArray";
@@ -14,6 +13,7 @@ import { MetadataMap } from "../../schemas/metadata/MetadataMap";
 import { MetadataObjectType } from "../../schemas/metadata/MetadataObjectType";
 import { MetadataSchema } from "../../schemas/metadata/MetadataSchema";
 import { MetadataSet } from "../../schemas/metadata/MetadataSet";
+import { MetadataStorage } from "../../schemas/metadata/MetadataStorage";
 import { MetadataTuple } from "../../schemas/metadata/MetadataTuple";
 import { MetadataTupleType } from "../../schemas/metadata/MetadataTupleType";
 import { IsProgrammer } from "../IsProgrammer";
